+++ .gitlab/release-md-write.py
@@ -13,6 +13,7 @@ vars = [
     'CI_COMMIT_SHORT_SHA',
     'GALACTEEK_VERSION',
     'APPIMAGE_FILENAME',
+    'APPIMAGE_DIR_CID',
     'APPIMAGE_IPFS_URL',
     'APPIMAGE_PATH',
     'FLATPAK_FILENAME',
@@ -24,25 +25,37 @@ vars = [
 prelude_ipfs_download = '''
 ## AppImage IPFS binary download
 
-This release is distributed on the IPFS network.
+This release is distributed on the IPFS network !
 
 You can [download it here]({url})
+
+If you run an IPFS node and want to help distributing this software, you
+can do so by run the following command:
+
+```
+ipfs pin add -r @@APPIMAGE_DIR_CID@@
+./@@APPIMAGE_FILENAME@@
+```
 '''
+
+
+def replace(msg):
+    for var in vars:
+        value = os.environ.get(var, None)
+        if value:
+            msg = msg.replace(
+                f"@@{var}@@",
+                value
+            )
+
+    return msg
 
 
 with open('.gitlab/RELEASE.md.tmpl', 'rt') as fd:
     md = fd.read()
 
     if appimage_url:
-        os.environ['PRELUDE'] = prelude_ipfs_download.format(
-            url=appimage_url)
+        os.environ['PRELUDE'] = replace(prelude_ipfs_download.format(
+            url=appimage_url))
 
-    for var in vars:
-        value = os.environ.get(var, None)
-        if value:
-            md = md.replace(
-                f"@@{var}@@",
-                value
-            )
-
-    print(md, file=sys.stdout)
+    print(replace(md), file=sys.stdout)